improve python3 compat

+++ pylsdj/project.py
@@ -1,8 +1,8 @@
 import sys, math
-import bread
+from bread import bread
 import bread_spec as spec
 from song import Song
-import StringIO
+import io
 import utils
 import filepack
 import blockutils
@@ -21,7 +21,7 @@ def load_lsdsng(filename):
 
     with open(filename, 'rb') as fp:
         # Skip the preamble this time around
-        fp.seek(len(preamble_data) / 8)
+        fp.seek(int(len(preamble_data) / 8))
 
         # Load compressed data into a block map and use BlockReader to
         # decompress it
